shenron/requirements.txt
Django==4.2
mysqlclient==2.1.1
python-dotenv==1.0.0
django-environ==0.9.0
whitenoise==6.4.0
mysql-connector-python==9.1.0
gunicorn==20.1.0
djangorestframework==3.14.0
requests
PyJWT<2.0.0
django-allauth>=0.50.0
cryptography
six